backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
supabase==2.3.0
psycopg2-binary==2.9.9

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Data Processing
python-dateutil==2.8.2
python-multipart==0.0.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx>=0.24.0

# Utilities
email-validator==2.1.0
